KSENROLL-14743 Setup fusion-maven-plugin configuration
In order for the pull request CI to work it needs to be able to use the
fusion plugin's TagMojo to change the pom versions to a pull request specific
version.

This adds in the configuration to support this which will be invoked by the
pull-request-builder run-builder.sh script.

Remove branch qualifier as its not needed

We run against the local db so we don't need a branch qualifier.